Added third full Serial Port to Sparky 2
On the Sparky 2 ServOut5 and ServoOutpin6 there is the UART4_TX and
UART4_RX available.

Changes were made to the Sparky2 board_hw_defs.c to specify the new port
and in pios_board.c the PIOS_HAL_ConfigurePort gets executed for the new
port.

The hwsparky2.xml uavobjectdefinition was changed to mirror the port_types
for the new port. These are the same as the Mainport